refactor: 검진 로직 hook으로 추상화
